pyproject.toml
[tool.poetry]
name = "experiments"
version = "0.1.0"
description = ""
authors = ["polymorphisma <shrawansunar.6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.49.1"
requests = "^2.32.3"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
prettytable = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
